=moved some parts of the codes handling matplotlib backend to a dedicated utils file
